Update TreasureMap.cs
Changed struct Point2d to class. Redefined equals for it. Added class Node to build a path. Added some methods to calculate the optimised path with a* algorithm. Does not work if try to filter points with river line. Does not choose optimised exit from base, does not optimised for case with two treasures.